Issue #186: Move IntentValidator.getUriToOpen to ext fn.
We can simplify the code by just calling extension functions instead of
wrapping the type.
