+++ app/src/test/java/org/mozilla/focus/ext/SafeIntentKtTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus
+package org.mozilla.focus.ext
 
 import android.content.Context
 import android.content.Intent
@@ -14,22 +14,20 @@ import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mozilla.focus.ext.toSafeIntent
-import org.mozilla.focus.ext.toUri
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
 
 private const val TEST_URL = "https://github.com/mozilla-mobile/focus-android"
 
 @RunWith(RobolectricTestRunner::class)
-class IntentValidatorTest {
+class SafeIntentKtTest {
 
     private val context: Context get() = RuntimeEnvironment.application
 
     @Test
     fun `WHEN receiving an intent with a null action THEN a null uri is returned`() {
         val intent = Intent(null, TEST_URL.toUri()).toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertNull(actual)
     }
 
@@ -37,7 +35,7 @@ class IntentValidatorTest {
     fun `WHEN receiving a view intent with a valid uri THEN the uri is returned`() {
         val expectedUrl = TEST_URL
         val intent = Intent(Intent.ACTION_VIEW, Uri.parse(expectedUrl)).toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertEquals(expectedUrl, actual)
     }
 
@@ -45,7 +43,7 @@ class IntentValidatorTest {
     @Test
     fun `WHEN receiving a view intent with a null uri THEN a null uri is returned`() {
         val intent = Intent(Intent.ACTION_VIEW, null).toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertNull(actual)
     }
 
@@ -56,7 +54,7 @@ class IntentValidatorTest {
             "     "
         ).forEachIndexed { i, blankStr ->
             val intent = Intent(Intent.ACTION_VIEW, blankStr.toUri()).toSafeIntent()
-            val actual = IntentValidator(intent).getUriToOpen(context)
+            val actual = intent.getUriToOpen(context)
             assertNull("index $i", actual)
         }
     }
@@ -67,13 +65,13 @@ class IntentValidatorTest {
         // to handle them specially and we handle them like view intents.
         val expectedUrl = TEST_URL
         val intent = CustomTabsIntent.Builder()
-                .setToolbarColor(Color.GREEN)
-                .addDefaultShareMenuItem()
-                .build()
-                .intent
-                .setData(Uri.parse(expectedUrl))
-                .toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+            .setToolbarColor(Color.GREEN)
+            .addDefaultShareMenuItem()
+            .build()
+            .intent
+            .setData(Uri.parse(expectedUrl))
+            .toSafeIntent()
+        val actual = intent.getUriToOpen(context)
         assertEquals(expectedUrl, actual)
     }
 
@@ -84,7 +82,7 @@ class IntentValidatorTest {
             addFlags(Intent.FLAG_ACTIVITY_LAUNCHED_FROM_HISTORY)
         }.toSafeIntent()
 
-        assertEquals(expectedUrl, IntentValidator(intent).getUriToOpen(context))
+        assertEquals(expectedUrl, intent.getUriToOpen(context))
     }
 
     @Test
@@ -93,7 +91,7 @@ class IntentValidatorTest {
         val intent = Intent(Intent.ACTION_SEND).apply {
             putExtra(Intent.EXTRA_TEXT, expectedUrl)
         }.toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertEquals(expectedUrl, actual)
     }
 
@@ -104,7 +102,7 @@ class IntentValidatorTest {
             putExtra(Intent.EXTRA_TEXT, expectedText)
         }.toSafeIntent()
 
-        val searchUrl = IntentValidator(intent).getUriToOpen(context)
+        val searchUrl = intent.getUriToOpen(context)
         expectedText.split(" ").forEach {
             assertTrue("Expected search url to contain $it", searchUrl!!.contains(it))
         }
@@ -113,7 +111,7 @@ class IntentValidatorTest {
     @Test
     fun `WHEN receiving a share intent with no text extra THEN a null uri is returned`() {
         val intent = Intent(Intent.ACTION_SEND).toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertNull(actual)
     }
 
@@ -122,7 +120,7 @@ class IntentValidatorTest {
         val intent = Intent(Intent.ACTION_SEND).apply {
             putExtra(Intent.EXTRA_TEXT, null as String?)
         }.toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertNull(actual)
     }
 
@@ -131,7 +129,7 @@ class IntentValidatorTest {
         val intent = Intent(Intent.ACTION_SEND).apply {
             putExtra(Intent.EXTRA_TEXT, "    ")
         }.toSafeIntent()
-        val actual = IntentValidator(intent).getUriToOpen(context)
+        val actual = intent.getUriToOpen(context)
         assertNull(actual)
     }
 
@@ -144,7 +142,7 @@ class IntentValidatorTest {
             Intent.ACTION_CALL
         ).forEach {
             val intent = Intent(it, TEST_URL.toUri()).toSafeIntent()
-            val actual = IntentValidator(intent).getUriToOpen(context)
+            val actual = intent.getUriToOpen(context)
             assertNull("Expeceted null for action $it", actual)
         }
     }
